certs: add reloadable `Certificate` type
This commit adds a new `Certificate` type
that supports reloading:
 - after a time period
 - changes to the certificate / private key file
 - OS signals

This type will be used by MinIO to implement
reloading a certificate - e.g. for KES, or some WebHook.
